feat: Info삭제&DBUser생성, getdatamethod
