--- LovelyCrane/Model/Info.swift
@@ -1,13 +1,0 @@
-//
-//  User.swift
-//  LoveCrane
-//
-//  Created by 235 on 2023/07/14.
-//
-
-import Foundation
-struct Info: Codable {
-    let nickname: String?
-    let partnerId: String?
-    let uuid: String?
-}
